+++ activejob/test/jobs/sanitized_param_job.rb
@@ -1,0 +1,11 @@
+class SanitizedParamJob < ActiveJob::Base
+  def perform(one, two)
+    logger.info 'Dummy'
+  end
+
+  def arguments_to_log
+    result = super.dup
+    result[1] = ActiveJob::Logging::SANITIZED_ARG
+    result
+  end
+end